categories/serializers/category_serializer.py
from rest_framework import serializers

from categories.models import Category


class CategorySerializer(serializers.ModelSerializer):
    id = serializers.IntegerField(required=False)

    class Meta:
        model = Category
        fields = '__all__'
