pyproject.toml
[tool.poetry]
name = "29"
version = "0.1.0"
description = ""
authors = ["example <svinokot@mail.mail>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.0.6"
Pillow = "^9.2.0"
djangorestframework = "^3.13.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
